Propagate parser errors to set exit code
The exit code of the application now depends on the parser errors. With
no errors, the exit code remains 0. If errors have been detected, the
exit code is set to 2 (to distinguish from IO errors, that have been
previously implemented to terminate the application immediately with an
exit code of 1).

The more elborate exit codes allow better use for automated processes
like CI.